Re-enable embargo and ensure all requests are made with a token
