+++ src/electron/frontend/core/components/pages/guided-mode/options/GuidedUpload.js
@@ -179,7 +179,7 @@ export class GuidedUploadPage extends Page {
                     },
                     onUpdate: () => (this.unsavedUpdates = true),
                     onThrow,
-                    validateOnChange: validate,
+                    validateOnChange: (...args) => validate.call(this, ...args)
                 }));
             })
             .catch((error) => html`<p>${error}</p>`);
