+++ src/schemas/dandi-create.schema.ts
@@ -1,4 +1,4 @@
-import create from './json/dandi/create_no_embargo.json' assert { type: "json" }
+import create from './json/dandi/create.json' assert { type: "json" }
 const schema = structuredClone(create)
 
 export default schema
